+++ miles/router/session/session_types.py
@@ -1,0 +1,15 @@
+from pydantic import BaseModel
+
+
+class SessionRecord(BaseModel):
+    timestamp: float
+    method: str
+    path: str
+    request: dict
+    response: dict
+    status_code: int
+
+
+class GetSessionResponse(BaseModel):
+    session_id: str
+    records: list[SessionRecord]
